combined keypress for all platforms
